keybase_service_base: don't for reload of cached public team info
We use cached team info when possible, unless we're checking for
membership of a user that's not present in the cached info.  In that
case, we force a reload, just in case there's a race with the
membership-changed notification from the service.  Usually this is
fine because it's rare to do a negative lookup multiple times in a row
-- after the first error, the client will stop.  However, in the case
of stat-ing all the files in a public directory, we call this function
over and over again to figure out if the current user is a writer (to
set the writer bit in each entry's permissions), and since it's a
public folder the user is not (usually) a writer, so it was forcing
the service to reload the team every time, slowing things down
considerably.

Instead, used the cached info in the case of public teams.  Their
membership can't change right now anyway, and if they ever do in the
future, it's probably a reasonable tradeoff.

Issue: HOTPOT-1009